kacmalpka-app/requirements.txt
Django==3.2.7
PyMySQL==1.1.0
django-bootstrap4==23.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
